requirement.txt
torch
transformers
diffusers
gradio
safetensors
huggingface_hub
Pillow
